requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# NLI models
datasets>=2.14.0

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0

# Graph processing
networkx>=3.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Evaluation
scikit-learn>=1.3.0
